feat(modal): fixed docs
