+++ docs/index.html
@@ -45,7 +45,6 @@
                 <li><a href="makeup-key-emitter/index.html">makeup-key-emitter</a></li>
                 <li><a href="makeup-keyboard-trap/index.html">makeup-keyboard-trap</a></li>
                 <li><a href="makeup-modal/index.html">makeup-modal</a></li>
-                <li><a href="makeup-modal-hoist/index.html">makeup-modal with hoisting</a></li>
                 <li><a href="makeup-navigation-emitter/index.html">makeup-navigation-emitter</a></li>
                 <li><a href="makeup-next-id/index.html">makeup-next-id</a></li>
                 <li><a href="makeup-prevent-scroll-keys/index.html">makeup-prevent-scroll-keys</a></li>
