--- docs/makeup-modal-hoist/index.html
@@ -1,56 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <title>makeup-modal demo</title>
-        <meta name="viewport" content="width=device-width, initial-scale=1"/>
-        <style>
-            .modal {
-                border: 1px dashed grey;
-                margin: 16px 0;
-                padding: 16px;
-            }
-        </style>
-    </head>
-    <body>
-        <main role="main">
-            <h1>makeup-modal demo</h1>
-            <p>Click on the button to apply the modal, with hoisting.</p>
-            <div>
-                <button>Hoist me!</button>
-            </div>
-            <div>
-                <button>Hoist #2!</button>
-            </div>
-            <div>
-                <button>Unhoist</button>
-            </div>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-            <a href="http://www.ebay.com">link</a>
-
-            <div class="modal">
-                <h2>Modal 1</h2>
-                <a href="http://www.ebay.com">link</a>
-                <input type="text" autofocus placeholder="textbox" />
-                <a href="http://www.ebay.com">link</a>
-            </div>
-
-            <a href="http://www.ebay.com">link</a>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-
-            <div class="modal">
-                <h2>Modal 2</h2>
-                <a href="http://www.ebay.com">link</a>
-                <input type="text" autofocus placeholder="textbox" />
-                <a href="http://www.ebay.com">link</a>
-            </div>
-
-            <a href="http://www.ebay.com">EBAY</a>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-
-            <a href="http://www.ebay.com">EBAY</a>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-            
-        </main>
-        <script src="index.min.js"></script>
-    </body>
-</html>
