+++ docs/makeup-modal/index.html
@@ -14,12 +14,27 @@
     <body>
         <main role="main">
             <h1>makeup-modal demo</h1>
-            <p>The dotted region below has been set to modal, using the modal() method.</p>
-
+            <p>Click on the button to apply the modal</p>
+            <div>
+                <button>Modal me!</button>
+            </div>
+            <div>
+                <button>Modal #2!</button>
+            </div>
+            <div>
+                <button>Unhoist</button>
+            </div>
+            <div>
+                <label for="hoist-checkbox">
+                    Check to hoist with modal!
+                </label>
+                <input id="hoist-checkbox" type="checkbox">
+            </div>
             <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
             <a href="http://www.ebay.com">link</a>
 
             <div class="modal">
+                <h2>Modal 1</h2>
                 <a href="http://www.ebay.com">link</a>
                 <input type="text" autofocus placeholder="textbox" />
                 <a href="http://www.ebay.com">link</a>
@@ -28,6 +43,19 @@
             <a href="http://www.ebay.com">link</a>
             <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
 
+            <div class="modal">
+                <h2>Modal 2</h2>
+                <a href="http://www.ebay.com">link</a>
+                <input type="text" autofocus placeholder="textbox" />
+                <a href="http://www.ebay.com">link</a>
+            </div>
+
+            <a href="http://www.ebay.com">EBAY</a>
+            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
+
+            <a href="http://www.ebay.com">EBAY</a>
+            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
+            
         </main>
         <script src="index.min.js"></script>
     </body>
